refactor: isolate option analytics client workflow
